refactor: 프로젝트 폴더 구조 수정 - admin / user 분리 - main 폴더 네이밍PascalCase 적용
