+++ src/pages/user/Main/api/club.ts
@@ -1,4 +1,4 @@
-import type { ClubResponse } from '@/pages/main/type/club';
+import type { ClubResponse } from '@/pages/user/Main/type/club';
 
 export async function getClubs(): Promise<ClubResponse> {
   const response = await fetch('/api/clubs');
