+++ src/pages/user/Main/components/BannerSection/BannerText.tsx
