+++ src/pages/user/Main/components/ClubCategorySection/index.tsx
@@ -1,14 +1,14 @@
-import { CLUB_CATEGORY } from '@/pages/main/constant/clubCategory';
+import { CLUB_CATEGORY } from '@/pages/user/Main/constant/clubCategory';
 import CategoryTab, {
   CategoryTabContainer,
-} from '@/pages/main/components/ClubCategorySection/CategoryTab';
+} from '@/pages/user/Main/components/ClubCategorySection/CategoryTab';
 
 const ClubCategorySection = () => {
   return (
     <>
       <CategoryTabContainer>
         {CLUB_CATEGORY.map((category) => (
-          <CategoryTab>{category}</CategoryTab>
+          <CategoryTab key={category}>{category}</CategoryTab>
         ))}
       </CategoryTabContainer>
     </>
