+++ src/pages/user/Main/hook/useClub.ts
@@ -1,6 +1,6 @@
-import { getClubs } from '@/pages/main/api/club';
+import { getClubs } from '@/pages/user/Main/api/club';
 import { useEffect, useState } from 'react';
-import type { Club } from '@/pages/main/type/club';
+import type { Club } from '@/pages/user/Main/type/club';
 
 export const useClub = () => {
   const [clubs, setClubs] = useState<Club[]>();
